+++ CHANGELOG.md
@@ -11,3 +11,5 @@
 #### core
 
 -   Added input word deletion support [#154](https://github.com/LaunchMenu/LaunchMenu/pull/154)
+-   Fixed syntax highlighting updating [#159](https://github.com/LaunchMenu/LaunchMenu/pull/159)
+-   Added standard item highlighting customization [#160](https://github.com/LaunchMenu/LaunchMenu/pull/160)